feat: query문 분리
